# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Biblioteca — API
  version: "1.0.0"
  description: |
    API para un sistema de gestión de biblioteca (Java 21, H2) que demuestra múltiples patrones de diseño.
    Esta especificación describe endpoints para agregar, buscar, listar y prestar libros.
    Implementaciones requeridas (no parte del contrato): Singleton (DB), Factory Method,
    Abstract Factory, Builder, Strategy, Observer, Decorator, Chain of Responsibility, Adapter.
servers:
  - url: http://localhost:8080
    description: Servidor local (ejemplo)
...
paths:
  /api/books:
    post:
      tags: [ Books ]
      summary: Agregar un libro
      description: |
        Agrega un libro a la biblioteca. Se aplican validaciones (Chain of Responsibility:
        ValidadorTitulo, ValidadorAutor). El servidor devolverá el libro creado.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBookRq'
      responses:
        "201":
          description: Libro creado correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateBookRs'
        "400":
          description: Validación fallida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
    get:
      tags: [ Books ]
      summary: Listar o buscar libros
      description: |
        Lista todos los libros o realiza búsqueda con diferentes estrategias.
        Parámetros de consulta:
          - ?title=...  (SearchByTitle)
          - ?author=... (SearchByAuthor)
        También se permite ?strategy=title|author para forzar la estrategia.
      parameters:
        - name: title
          in: query
          schema:
            type: string
          description: Buscar por título (parte o completo)
        - name: author
          in: query
          schema:
            type: string
          description: Buscar por autor (parte o completo)
        #        - name: strategy
        #          in: query
        #          schema:
        #            type: string
        #            enum: [title, author, smart]
        #            default: smart
        #          description: Estrategia de búsqueda (Strategy pattern). 'smart' = heurística que decide.
        - name: id
          in: query
          schema:
            type: string
          description: Buscar por UUID exacto
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Número de página (paginación simple)
        - name: size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Tamaño de página
      responses:
        "200":
          description: Lista de libros (posible conjunto vacío)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryBooksRs'
        "400":
          description: Parámetros inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    CreateBookRq:
      type: object
      description: Entidad para crear un libro
      properties:
        book:
          required:
            - title
            - author
            - type
            - format
          allOf:
            - $ref: '#/components/schemas/Book'
    CreateBookRs:
      type: object
      description: Respuesta para creación de libro
      properties:
        book:
          $ref: '#/components/schemas/Book'
    QueryBooksRs:
      type: object
      description: Respuesta para consulta de libros
      properties:
        books:
          type: array
          items:
            $ref: '#/components/schemas/Book'
...
    Book:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: UUID generado por el sistema
        title:
          type: string
          description: Título del libro
        author:
          type: string
          description: Nombre del autor o autores
        type:
          type: string
          description: Tipo de libro (Ficción/No Ficción)
          enum:
            - FICTION
            - NON-FICTION
        format:
          type: string
          description: Formato del libro (Físico/Digital)
          enum:
            - E-BOOK
            - PHYSICAL
        status:
          type: string
          description: Estado del libro en la biblioteca
          enum:
            - AVAILABLE
            - LOANED
        metadata:
          type: object
          description: Campos adicionales (p. ej. páginas, isbn, ubicacion, urlDescarga)
          additionalProperties: true

    # Payload para crear libro (Builder + Factory + Chain validators aplicados en servidor)
...
    ValidationError:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            type: string
      example:
        code: "VALIDATION_FAILED"
        message: "Título no puede estar vacío"
        details:
          - "titulo: valor vacío"
          - "autor: formato inválido"
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      example:
        code: "NOT_FOUND"
        message: "Libro no encontrado"
...